Pickup system finished
2 interfaces are used
IPickup is used for generic pickup tasks like add remove component events
Char is responsible for removing components, as it is the owner.

IpickupStatsUpdata is used to access stats in the char so the pickups can do there job.

base pick up is a generic pick up that is responsible for adding pickup components to the char.

there are 2 types of pickup component. immediate and timed.
immediate applies the effects immediately and removes its self.
timed has an effect until the timed function, Destructor< is called.

all pickup components have 3 methods, to be overridden by children as needed:
setup runs once when the component is added.
action run as the dev feels nessacery, on setup or on tick.
Destructor is used to clean-up reset char properties as needed

pickup base allows you to set the visual mess for the pickup, its color and the pickupcomponent to be added.